app/requirements.txt
psycopg2-binary
SQLAlchemy
sqlalchemy
flask_sqlalchemy
django
pymysql
djangorestframework
djangorestframework-simplejwt
django-cors-headers